improve: player를 딜러와 참가자 상속하도록 구조 수정
