+++ src/main/kotlin/blackject/model/GameResult.kt
@@ -3,4 +3,4 @@ package blackject.model
 /**
  * 게임 결과
  * */
-class GameResult(val result: HashMap<Person, Int>)
+class GameResult(val result: HashMap<Player, Int>)
