--- src/main/kotlin/blackject/model/PersonType.kt
@@ -1,8 +1,0 @@
-package blackject.model
-
-/**
- * 참가자 유형
- * */
-enum class PersonType {
-    DEALER, NORMAL
-}
